Add logged in signalling capability
This would allow a custom user of this application to certain events.
The one I need now is when a user is logged in. We build a custom
version of gavel-ci for ci.foundries.io that has some additional
user management functionality. This lets our app know when a user is
logged in.

Signed-off-by: Andy Doan <andy@foundries.io>