jobserv: Discard gzip related headers
We've recently put our API behind a cloudflare proxy which will compress
the response to capable clients like python-requests. However,
python-requests will leave certain headers related to the compression in
the response (while decompressing what you get for you). We need to
discard these so that we don't confuse the browser.

Signed-off-by: Andy Doan <andy@foundries.io>